orders/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-import-export==3.0.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
kombu==5.2.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.0
prompt-toolkit==3.0.36
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
ujson==5.7.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
